+++ controllers/login_controller.js
@@ -1,59 +1,15 @@
-const bcrypt = require('bcrypt');
-const dbHandler = require('../data/dbHandler');
-const {constructResponse} = require('../utils/utils.js');
+const loginService = require('../services/loginService');
+const dbHandler = require("../data/dbHandler");
 
-const unSuspend= async (user) => {
-    await dbHandler.updateUser(user.email, {"status": "active", "suspensionTime": 0, "suspensionDate": 0});
-}
 
-async function isSuspend(user) {
-    if (user.status === 'active') {
-        console.log(`user: ${user["email"]} is not suspended`);
-        return 0;
-    } else if (user.status === 'suspended') {
-        const today = new Date();
-        const suspendTime = parseInt(user.suspensionTime);
-        const suspendStartDate = user.suspensionDate;
-        let dateExpired = suspendStartDate;
-        dateExpired.setDate(suspendStartDate.getDate() + suspendTime)
-        const isSuspend = isAfter(dateExpired, today);
-        if (isSuspend) {
-            console.log(`user: ${user["email"]} is suspended- login failed`, 'ERROR');
-            return suspendStartDate + parseInt(suspendTime);
-        } else {
-            await unSuspend(user);
-            console.log(`user: ${user["email"]} is no longer suspended`);
-            return 0;
-        }
-    } else if (user.status === 'closed') {
-        console.log(`user: ${user["email"]} is closed forever! bye bye`);
-        return "forever";
+const loginControl = async function (req, res, next) {
+    try {
+        await loginService.handleLogin(req, res, next);
+        return res.sendStatus(200);
+    } catch (e) {
+        return res.status(401).json({message: e.message});
     }
 }
-
-const handleLogin = async (req, res, next) => {
-    const userEmail = req.body.email.toLowerCase();
-    const userPassword = req.body.password;
-    let user;
-
-try {
-    user = await dbHandler.getUserByEmail(userEmail) //maybe needs await in the start and in the end.lean()
-    if(!user){
-        throw new Error("user does'nt exist")
-    }
-    const suspend = await isSuspend(user);
-    if(suspend){
-        return res.status(403).json({message:`User is suspended until ${isSuspend}`});}
-    if (!await bcrypt.compare(userPassword, user.password)) { throw new Error("incorrect password")}
-        console.log(`password correct! ${user.email} welcome`);
-        const today=new Date();
-        await dbHandler.updateUser(userEmail, {"loginDate": today})
-        return res.status(200).json({message:"ok"})
-
-}catch(e) {
-    return res.status(401).json({message:e.message});
-}}
-
 
 const Permissions = async (req, res, next) => {
     try {
@@ -69,8 +25,5 @@ const Permissions = async (req, res, next) => {
     }
 }
 
-function isAfter(date1, date2) {
-    return date1 > date2;
-}
 
-module.exports = {handleLogin, Permissions}
+module.exports={Permissions,loginControl};
