+++ routers/logs.log
@@ -324,3 +324,94 @@
  GET /style.css 304 - - 5.640 ms Wed, 07 Dec 2022 18:04:49 GMT
  GET /script.js 200 17212 - 7.416 ms Wed, 07 Dec 2022 18:04:49 GMT
  GET /api/admin/showUser 200 883 - 73.778 ms Wed, 07 Dec 2022 18:04:51 GMT
+ GET / 200 5131 - 29.600 ms Wed, 07 Dec 2022 18:11:33 GMT
+ GET /style.css 200 11704 - 10.880 ms Wed, 07 Dec 2022 18:11:33 GMT
+ GET /script.js 200 17212 - 6.554 ms Wed, 07 Dec 2022 18:11:34 GMT
+ GET /BAMD%20IAM(1).png 200 53648 - 6.885 ms Wed, 07 Dec 2022 18:11:34 GMT
+ GET /favicon.ico 200 1437 - 6.874 ms Wed, 07 Dec 2022 18:11:34 GMT
+ POST /api/login 200 16 - 441.298 ms Wed, 07 Dec 2022 18:11:40 GMT
+ GET /homePage.html 200 6468 - 1.835 ms Wed, 07 Dec 2022 18:11:40 GMT
+ GET /homePage.html 304 - - 2.862 ms Wed, 07 Dec 2022 18:11:40 GMT
+ GET /style.css 304 - - 4.353 ms Wed, 07 Dec 2022 18:11:41 GMT
+ GET /script.js 304 - - 2.648 ms Wed, 07 Dec 2022 18:11:41 GMT
+ GET /index.html 200 5131 - 34.116 ms Wed, 07 Dec 2022 18:22:21 GMT
+ GET /style.css 304 - - 4.793 ms Wed, 07 Dec 2022 18:22:21 GMT
+ GET /script.js 304 - - 5.783 ms Wed, 07 Dec 2022 18:22:21 GMT
+ GET /BAMD%20IAM(1).png 304 - - 3.910 ms Wed, 07 Dec 2022 18:22:21 GMT
+ GET /index.html? 200 5131 - 5.000 ms Wed, 07 Dec 2022 18:22:27 GMT
+ GET /style.css 304 - - 6.247 ms Wed, 07 Dec 2022 18:22:27 GMT
+ GET /BAMD%20IAM(1).png 304 - - 4.901 ms Wed, 07 Dec 2022 18:22:27 GMT
+ GET /script.js 304 - - 3.872 ms Wed, 07 Dec 2022 18:22:27 GMT
+ GET /index.html? 304 - - 4.363 ms Wed, 07 Dec 2022 18:23:23 GMT
+ POST /api/login - - - - ms Wed, 07 Dec 2022 18:23:23 GMT
+ POST /api/login - - - - ms Wed, 07 Dec 2022 18:23:23 GMT
+ POST /api/login - - - - ms Wed, 07 Dec 2022 18:23:23 GMT
+ GET /style.css 304 - - 4.880 ms Wed, 07 Dec 2022 18:23:23 GMT
+ GET /BAMD%20IAM(1).png 304 - - 3.143 ms Wed, 07 Dec 2022 18:23:23 GMT
+ GET /script.js 304 - - 5.514 ms Wed, 07 Dec 2022 18:23:23 GMT
+ GET /favicon.ico 304 - - 17.729 ms Wed, 07 Dec 2022 18:23:23 GMT
+ GET /index.html? 200 5131 - 3.915 ms Wed, 07 Dec 2022 18:23:32 GMT
+ POST /api/login - - - - ms Wed, 07 Dec 2022 18:23:32 GMT
+ GET /style.css 200 11704 - 6.579 ms Wed, 07 Dec 2022 18:23:32 GMT
+ GET /script.js 200 17212 - 10.720 ms Wed, 07 Dec 2022 18:23:32 GMT
+ GET /BAMD%20IAM(1).png 200 53648 - 14.421 ms Wed, 07 Dec 2022 18:23:32 GMT
+ GET /favicon.ico 200 1437 - 4.691 ms Wed, 07 Dec 2022 18:23:32 GMT
+ GET /index.html? 304 - - 33.526 ms Wed, 07 Dec 2022 18:27:56 GMT
+ GET /style.css 304 - - 15.414 ms Wed, 07 Dec 2022 18:27:56 GMT
+ GET /script.js 304 - - 12.323 ms Wed, 07 Dec 2022 18:27:56 GMT
+ GET /BAMD%20IAM(1).png 304 - - 8.460 ms Wed, 07 Dec 2022 18:27:56 GMT
+ GET /favicon.ico 304 - - 8.493 ms Wed, 07 Dec 2022 18:27:56 GMT
+ POST /api/login 200 2 - 1164.841 ms Wed, 07 Dec 2022 18:28:06 GMT
+ GET /homePage.html 200 6468 - 14.927 ms Wed, 07 Dec 2022 18:28:06 GMT
+ GET /style.css 304 - - 6.185 ms Wed, 07 Dec 2022 18:28:06 GMT
+ GET /script.js 304 - - 2.965 ms Wed, 07 Dec 2022 18:28:06 GMT
+ GET /style.css 304 - - 24.730 ms Wed, 07 Dec 2022 18:28:43 GMT
+ GET /index.html? 200 5131 - 11.443 ms Wed, 07 Dec 2022 18:28:44 GMT
+ GET /style.css 200 11704 - 15.065 ms Wed, 07 Dec 2022 18:28:44 GMT
+ GET /script.js 200 17212 - 7.448 ms Wed, 07 Dec 2022 18:28:44 GMT
+ GET /BAMD%20IAM(1).png 200 53648 - 7.194 ms Wed, 07 Dec 2022 18:28:44 GMT
+ GET /favicon.ico 200 1437 - 32.893 ms Wed, 07 Dec 2022 18:28:44 GMT
+ POST /api/login 401 35 - 491.807 ms Wed, 07 Dec 2022 18:28:50 GMT
+ POST /api/login 200 16 - 1494.939 ms Wed, 07 Dec 2022 18:31:12 GMT
+ GET /homePage.html 200 6468 - 13.981 ms Wed, 07 Dec 2022 18:31:12 GMT
+ GET /style.css 304 - - 6.867 ms Wed, 07 Dec 2022 18:31:14 GMT
+ GET /script.js 304 - - 5.830 ms Wed, 07 Dec 2022 18:31:14 GMT
+ GET /homePage.html 200 6468 - 12.262 ms Wed, 07 Dec 2022 18:31:56 GMT
+ GET /style.css 200 11704 - 51.369 ms Wed, 07 Dec 2022 18:31:57 GMT
+ GET /script.js 200 17212 - 10.070 ms Wed, 07 Dec 2022 18:31:57 GMT
+ GET /favicon.ico 200 1437 - 17.165 ms Wed, 07 Dec 2022 18:31:58 GMT
+ GET /style.css 304 - - 3.020 ms Wed, 07 Dec 2022 18:31:59 GMT
+ POST /api/login 200 16 - 734.508 ms Wed, 07 Dec 2022 18:32:01 GMT
+ GET /homePage.html 200 6468 - 5.288 ms Wed, 07 Dec 2022 18:32:01 GMT
+ GET /style.css 304 - - 1.476 ms Wed, 07 Dec 2022 18:32:03 GMT
+ GET /script.js 304 - - 3.074 ms Wed, 07 Dec 2022 18:32:03 GMT
+ POST /api/login 200 2 - 793.913 ms Wed, 07 Dec 2022 18:32:19 GMT
+ GET /homePage.html 200 6468 - 12.555 ms Wed, 07 Dec 2022 18:32:19 GMT
+ GET /style.css 304 - - 5.424 ms Wed, 07 Dec 2022 18:32:19 GMT
+ GET /script.js 304 - - 4.123 ms Wed, 07 Dec 2022 18:32:19 GMT
+ GET /homePage.html 200 6468 - 30.393 ms Wed, 07 Dec 2022 18:56:42 GMT
+ GET /style.css 200 11704 - 10.088 ms Wed, 07 Dec 2022 18:56:43 GMT
+ GET /script.js 200 17212 - 15.135 ms Wed, 07 Dec 2022 18:56:43 GMT
+ GET /favicon.ico 200 1437 - 3.975 ms Wed, 07 Dec 2022 18:56:44 GMT
+ GET / 200 5131 - 7.236 ms Wed, 07 Dec 2022 18:56:47 GMT
+ GET /style.css 304 - - 15.277 ms Wed, 07 Dec 2022 18:56:47 GMT
+ GET /script.js 304 - - 6.174 ms Wed, 07 Dec 2022 18:56:47 GMT
+ GET /BAMD%20IAM(1).png 200 53648 - 13.112 ms Wed, 07 Dec 2022 18:56:47 GMT
+ POST /api/login 200 2 - 504.412 ms Wed, 07 Dec 2022 18:56:52 GMT
+ GET /homePage.html 304 - - 4.507 ms Wed, 07 Dec 2022 18:56:52 GMT
+ GET /homePage.html 304 - - 2.833 ms Wed, 07 Dec 2022 18:56:52 GMT
+ GET /style.css 304 - - 4.836 ms Wed, 07 Dec 2022 18:56:53 GMT
+ GET /script.js 304 - - 2.811 ms Wed, 07 Dec 2022 18:56:53 GMT
+ GET /style.css 304 - - 3.636 ms Wed, 07 Dec 2022 18:56:56 GMT
+ POST /api/login 401 32 - 700.967 ms Wed, 07 Dec 2022 18:57:00 GMT
+ POST /api/login 401 32 - 248.493 ms Wed, 07 Dec 2022 18:57:11 GMT
+ POST /api/login 401 32 - 82.188 ms Wed, 07 Dec 2022 18:57:17 GMT
+ GET /? 200 5131 - 3.005 ms Wed, 07 Dec 2022 18:57:25 GMT
+ GET /style.css 304 - - 3.631 ms Wed, 07 Dec 2022 18:57:25 GMT
+ GET /BAMD%20IAM(1).png 304 - - 3.146 ms Wed, 07 Dec 2022 18:57:25 GMT
+ GET /script.js 304 - - 6.799 ms Wed, 07 Dec 2022 18:57:25 GMT
+ POST /api/login 200 2 - 228.008 ms Wed, 07 Dec 2022 18:57:33 GMT
+ GET /homePage.html - - - - ms Wed, 07 Dec 2022 18:57:33 GMT
+ GET /homePage.html 200 6468 - 4.829 ms Wed, 07 Dec 2022 18:57:33 GMT
+ GET /style.css 304 - - 1.945 ms Wed, 07 Dec 2022 18:57:33 GMT
+ GET /script.js 304 - - 1.642 ms Wed, 07 Dec 2022 18:57:33 GMT
