--- utils/utils.js
@@ -1,8 +1,0 @@
-exports.ErrorResponse = function ErrorResponse(
-    response,
-    message,
-    statusCode
-) {
-    response.status(statusCode).json({ message: message});
-};
-
